bot_gui/requirements.txt
aiogram==3.10
python-dotenv==1.0